+++ testing/base/BUILD
@@ -82,8 +82,8 @@ cc_test(
         ":source_gen_lib",
         "//common:all_llvm_targets",
         "//common:set",
+        "//toolchain/base:install_paths_test_helpers",
         "//toolchain/driver",
-        "//toolchain/install:install_paths_test_helpers",
         "@googletest//:gtest",
         "@llvm-project//llvm:Support",
     ],
