+++ toolchain/driver/BUILD
@@ -38,9 +38,9 @@ cc_library(
         "//common:ostream",
         "//common:vlog",
         "//third_party/llvm:clang_cc1",
+        "//toolchain/base:install_paths",
         "//toolchain/base:kind_switch",
         "//toolchain/base:runtime_sources",
-        "//toolchain/install:install_paths",
         "@llvm-project//clang:basic",
         "@llvm-project//clang:clang-driver",
         "@llvm-project//clang:codegen",
@@ -72,7 +72,7 @@ cc_test(
         "//testing/base:file_helpers",
         "//testing/base:global_exe_path",
         "//testing/base:gtest_main",
-        "//toolchain/install:install_paths",
+        "//toolchain/base:install_paths",
         "@googletest//:gtest",
         "@llvm-project//llvm:Object",
         "@llvm-project//llvm:Support",
@@ -97,8 +97,8 @@ cc_test(
         "//testing/base:file_helpers",
         "//testing/base:global_exe_path",
         "//testing/base:gtest_main",
+        "//toolchain/base:install_paths",
         "//toolchain/base:llvm_tools",
-        "//toolchain/install:install_paths",
         "@googletest//:gtest",
         "@llvm-project//llvm:Object",
         "@llvm-project//llvm:Support",
@@ -116,7 +116,7 @@ cc_binary(
         "//testing/base:benchmark_main",
         "//testing/base:global_exe_path",
         "//testing/base:source_gen_lib",
-        "//toolchain/install:install_paths_test_helpers",
+        "//toolchain/base:install_paths_test_helpers",
         "//toolchain/testing:compile_helper",
         "@google_benchmark//:benchmark",
         "@llvm-project//llvm:Support",
@@ -181,6 +181,7 @@ cc_library(
         "//common:version",
         "//common:vlog",
         "//toolchain/base:clang_invocation",
+        "//toolchain/base:install_paths",
         "//toolchain/base:llvm_tools",
         "//toolchain/base:shared_value_stores",
         "//toolchain/base:timings",
@@ -189,7 +190,6 @@ cc_library(
         "//toolchain/diagnostics:diagnostic_emitter",
         "//toolchain/diagnostics:sorting_diagnostic_consumer",
         "//toolchain/format",
-        "//toolchain/install:install_paths",
         "//toolchain/language_server",
         "//toolchain/lex",
         "//toolchain/lower",
@@ -218,8 +218,8 @@ cc_test(
         "//testing/base:file_helpers",
         "//testing/base:global_exe_path",
         "//testing/base:gtest_main",
+        "//toolchain/base:install_paths",
         "//toolchain/diagnostics:diagnostic_emitter",
-        "//toolchain/install:install_paths",
         "//toolchain/lex:tokenized_buffer_test_helpers",
         "//toolchain/testing:yaml_test_helpers",
         "@googletest//:gtest",
@@ -238,7 +238,7 @@ cc_fuzz_test(
         "//common:exe_path",
         "//common:raw_string_ostream",
         "//testing/fuzzing:libfuzzer_header",
-        "//toolchain/install:install_paths",
+        "//toolchain/base:install_paths",
         "@llvm-project//llvm:Support",
     ],
 )
@@ -251,7 +251,7 @@ cc_library(
         ":tool_runner_base",
         "//common:ostream",
         "//common:vlog",
-        "//toolchain/install:install_paths",
+        "//toolchain/base:install_paths",
         "@llvm-project//lld:Common",
         "@llvm-project//lld:ELF",
         "@llvm-project//lld:MachO",
@@ -289,8 +289,8 @@ cc_library(
         ":tool_runner_base",
         "//common:ostream",
         "//common:vlog",
+        "//toolchain/base:install_paths",
         "//toolchain/base:llvm_tools",
-        "//toolchain/install:install_paths",
         "@llvm-project//lld:Common",
         "@llvm-project//lld:ELF",
         "@llvm-project//lld:MachO",
@@ -326,7 +326,7 @@ cc_library(
         "//common:ostream",
         "//common:version",
         "//common:vlog",
-        "//toolchain/install:install_paths",
+        "//toolchain/base:install_paths",
         "@llvm-project//llvm:Support",
     ],
 )
@@ -363,7 +363,7 @@ cc_library(
     deps = [
         "//common:ostream",
         "//common:vlog",
-        "//toolchain/install:install_paths",
+        "//toolchain/base:install_paths",
         "@llvm-project//llvm:Support",
     ],
 )
