+++ toolchain/driver/compile_benchmark.cpp
@@ -9,8 +9,8 @@
 
 #include "testing/base/global_exe_path.h"
 #include "testing/base/source_gen.h"
+#include "toolchain/base/install_paths_test_helpers.h"
 #include "toolchain/driver/driver.h"
-#include "toolchain/install/install_paths_test_helpers.h"
 #include "toolchain/testing/compile_helper.h"
 
 namespace Carbon::Testing {
