+++ toolchain/driver/driver_env.h
@@ -12,9 +12,9 @@
 #include "llvm/Support/ThreadPool.h"
 #include "llvm/Support/Threading.h"
 #include "llvm/Support/VirtualFileSystem.h"
+#include "toolchain/base/install_paths.h"
 #include "toolchain/diagnostics/diagnostic_emitter.h"
 #include "toolchain/driver/runtimes_cache.h"
-#include "toolchain/install/install_paths.h"
 
 namespace Carbon {
