+++ toolchain/language_server/BUILD
@@ -21,8 +21,8 @@ cc_library(
         ":outgoing_messages",
         "//common:ostream",
         "//common:raw_string_ostream",
+        "//toolchain/base:install_paths",
         "//toolchain/diagnostics:diagnostic_emitter",
-        "//toolchain/install:install_paths",
         "@llvm-project//clang-tools-extra/clangd:ClangDaemon",
     ],
 )
@@ -36,11 +36,11 @@ cc_library(
         "//common:map",
         "//common:raw_string_ostream",
         "//toolchain/base:clang_invocation",
+        "//toolchain/base:install_paths",
         "//toolchain/base:shared_value_stores",
         "//toolchain/check",
         "//toolchain/diagnostics:diagnostic_emitter",
         "//toolchain/diagnostics:file_diagnostics",
-        "//toolchain/install:install_paths",
         "//toolchain/lex",
         "//toolchain/lex:tokenized_buffer",
         "//toolchain/parse",
